+++ kod/radixsort.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 def extract_key(start_cifra, end_cifra, klic):
     """Vrátí část klíče od start_cifra do end_cifra
        >>> extract_key(1,4,12345)
@@ -6,7 +5,7 @@ def extract_key(start_cifra, end_cifra, klic):
        >>> extract_key(0,2,123445)
        45
     """
-    return (klic % (10**end_cifra))/(10**start_cifra)
+    return int((klic % (10**end_cifra))/(10**start_cifra))
 
 def var_buck_sort(seznam, scf, ecf):
     """ Provede bucket sort ale místo všech cifer klíče uvažuje pouze
@@ -43,7 +42,7 @@ from math import log, ceil
 def radix_sort(seznam, d):
     mx = max(seznam)
     delka_klice = int(ceil(log(mx,10)))
-    delka_casti = delka_klice/d
+    delka_casti = int(delka_klice/d)
     ecf = delka_casti
     while ecf <= delka_klice:
         var_buck_sort(seznam, ecf-delka_casti, ecf)
